完成基本的FlowField寻路，完成调试界面

+++ Assets/Scripts/ECS/Component/MapCollidersInfo.cs
@@ -49,7 +49,7 @@ public struct MapColliderInfo
 {
     public byte CostField;
 
-    public float IntegrationField;
+    public int IntegrationField;
 
     public IntegrateFlag IntegrateInfo;
